test: split review policy tests
