+++ test/clawsweeper.test.ts
@@ -1,5 +1,4 @@
 import assert from "node:assert/strict";
-import { execFileSync } from "node:child_process";
 import { existsSync, mkdirSync, mkdtempSync, readFileSync, rmSync, writeFileSync } from "node:fs";
 import { join } from "node:path";
 import test from "node:test";
@@ -10,66 +9,40 @@ import {
   auditHasStrictFailures,
   auditHealthSection,
   closingPullRequestReferenceTarget,
-  compactPullRequestForTest,
   codexEnv,
   codexLoginConfig,
   codexLoginMethod,
   dashboardClosedAt,
-  failedReviewRetryEligibilityForTest,
-  fixedPullRequestFromCommitPullsForTest,
   formatRecentClosedRows,
   ghRetryKind,
   ghRetryWaitMs,
-  isInfrastructureFailedReviewForTest,
   isGitHubNotFoundError,
   isGitHubRequiresAuthenticationError,
   isLockedConversationCommentError,
-  isProtectedItem,
   itemNumbersArg,
   lockedConversationApplyReason,
-  parseGhJson,
-  parseGhJsonLines,
   parseDecision,
-  protectedLabels,
   relatedGitHubIssueSearchQueryForTest,
   relatedTitleSearchTerms,
   renderReviewStartStatusComment,
   reviewArtifactDestination,
-  reviewAutomationMarkersFromReport,
-  reviewActionForDecision,
   reviewCodexForcedLoginMethodForTest,
-  reviewPromptForTest,
-  renderReviewCommentFromReport,
-  renderReviewContextBudgetForTest,
-  reviewContextLedgerForTest,
-  reviewDecisionSchemaText,
-  reviewPromptTelemetryForTest,
-  reviewPromptTemplate,
   runtimeBudgetExceeded,
   safeOutputTail,
   shardItemNumbers,
   shouldSyncReviewComment,
   shouldRetryGh,
-  shouldPlanItem,
-  validateCloseDecision,
 } from "../dist/clawsweeper.js";
-import { checkConclusionForFrontMatter } from "../dist/commit-checks.js";
-import { skippedNonCodeReport } from "../dist/commit-classifier.js";
-import {
-  commitReportRelativePath,
-  isReviewableCommitPath,
-  parseCommitReportSince,
-  parseCoAuthors,
-} from "../dist/commit-sweeper.js";
 import { parseArgs as parseClawsweeperArgs } from "../dist/clawsweeper-args.js";
 import { AUTOMATION_LIMITS } from "../dist/limits.js";
 import {
+  auditRecord,
   closeDecision,
-  git,
   implementedCloseReport,
   item,
   lowSignalCloseReport,
   promotionGhMock,
+  reportFrontMatter,
   reportWithSyncedReviewComment,
   runApplyDecisionsForTest,
   stalePullRequestReport,
@@ -78,1013 +51,6 @@ import {
   withMockGh,
   workPlanCandidateReport,
 } from "./helpers.ts";
-
-test("review prompt assets match tracked files", () => {
-  assert.equal(reviewPromptTemplate(), readFileSync("prompts/review-item.md", "utf8"));
-  assert.deepEqual(
-    JSON.parse(reviewDecisionSchemaText()),
-    JSON.parse(readFileSync("schema/clawsweeper-decision.schema.json", "utf8")),
-  );
-});
-
-test("sweep apply jobs wire the default-off product direction policy gate", () => {
-  const workflow = readFileSync(".github/workflows/sweep.yml", "utf8");
-  assert.equal(
-    workflow.match(/CLAWSWEEPER_UNCONFIRMED_PRODUCT_DIRECTION_CLOSE_ENABLED:/g)?.length,
-    2,
-  );
-  assert.match(
-    workflow,
-    /vars\.CLAWSWEEPER_UNCONFIRMED_PRODUCT_DIRECTION_CLOSE_ENABLED \|\| 'false'/,
-  );
-});
-
-test("main CLI args ignore package-manager double dash separators", () => {
-  assert.deepEqual(parseClawsweeperArgs(["apply-decisions", "--", "--dry-run"]), {
-    _: ["apply-decisions"],
-    dry_run: true,
-  });
-  assert.deepEqual(parseClawsweeperArgs(["apply-decisions", "--limit", "1", "--", "--dry-run"]), {
-    _: ["apply-decisions"],
-    limit: "1",
-    dry_run: true,
-  });
-});
-
-function reportFrontMatter(overrides = {}) {
-  const values = {
-    repository: "openclaw/openclaw",
-    type: "issue",
-    decision: "keep_open",
-    close_reason: "none",
-    confidence: "high",
-    action_taken: "kept_open",
-    ...overrides,
-  };
-  return `---
-${Object.entries(values)
-  .map(([key, value]) => `${key}: ${value}`)
-  .join("\n")}
----
-`;
-}
-
-function detailsBody(markdown, summary) {
-  const marker = `<summary>${summary}</summary>`;
-  const markerIndex = markdown.indexOf(marker);
-  assert.notEqual(markerIndex, -1, `missing details summary ${summary}`);
-  const bodyStart = markerIndex + marker.length;
-  const bodyEnd = markdown.indexOf("</details>", bodyStart);
-  assert.notEqual(bodyEnd, -1, `missing details close for ${summary}`);
-  return markdown.slice(bodyStart, bodyEnd);
-}
-
-function auditRecord(number, overrides = {}) {
-  return {
-    repo: "openclaw/openclaw",
-    number,
-    location: "items",
-    path: `items/${number}.md`,
-    kind: "issue",
-    title: `Item ${number}`,
-    labels: [],
-    decision: "keep_open",
-    closeReason: "none",
-    action: "kept_open",
-    reviewStatus: "complete",
-    currentState: undefined,
-    ...overrides,
-  };
-}
-
-test("review prompt telemetry records durable cost proxies", () => {
-  const context = {
-    issue: { number: 123, title: "Sample item" },
-    comments: [{ author: "contributor", body: "This still reproduces." }],
-    timeline: [],
-    counts: { comments: 1, timeline: 0 },
-  };
-
-  const telemetry = reviewPromptTelemetryForTest(
-    item({ title: "Telemetry regression" }),
-    context,
-    git,
-    "keep extra instructions visible",
-  );
-
-  assert.ok(telemetry.staticPromptChars > 1000);
-  assert.ok(telemetry.schemaChars > 1000);
-  assert.ok(telemetry.contextChars >= JSON.stringify(context, null, 2).length);
-  assert.ok(telemetry.promptChars > telemetry.staticPromptChars + telemetry.contextChars);
-  assert.equal(telemetry.additionalPromptChars, "keep extra instructions visible".length);
-});
-
-test("review prompt includes compact previous review state without raw durable review body", () => {
-  const context = {
-    issue: { number: 123, title: "Sample PR" },
-    comments: [{ author: "contributor", body: "After-fix proof is attached." }],
-    timeline: [],
-    previousClawSweeperReview: {
-      status: "found issues before merge.",
-      reviewedSha: "abc123",
-      summary: "Prior review found one blocker.",
-    },
-    counts: { comments: 3, commentsIncluded: 1, commentsFiltered: 2, timeline: 0 },
-  };
-
-  const prompt = reviewPromptForTest(item({ kind: "pull_request", number: 123 }), context, git);
-
-  assert.match(prompt, /"previousClawSweeperReview"/);
-  assert.match(prompt, /Prior review found one blocker/);
-  assert.match(prompt, /"commentsFiltered": 2/);
-  assert.doesNotMatch(prompt, /How this review workflow works/);
-});
-
-test("review prompt includes merge state and guards clean behind-branch drift", () => {
-  const compactPullRequest = compactPullRequestForTest({
-    number: 123,
-    title: "Sample PR",
-    html_url: "https://github.com/openclaw/openclaw/pull/123",
-    state: "open",
-    draft: false,
-    merged: false,
-    mergeable: true,
-    mergeable_state: "clean",
-    head: { ref: "feature", sha: "head123" },
-    base: { ref: "main", sha: "base123" },
-    user: { login: "contributor" },
-    additions: 10,
-    deletions: 2,
-    changed_files: 1,
-  });
-  const context = {
-    issue: { number: 123, title: "Sample PR" },
-    comments: [],
-    timeline: [],
-    pullRequest: compactPullRequest,
-    counts: { comments: 0, timeline: 0 },
-  };
-
-  const prompt = reviewPromptForTest(item({ kind: "pull_request", number: 123 }), context, git);
-
-  assert.deepEqual((compactPullRequest as { mergeableState?: unknown }).mergeableState, "clean");
-  assert.match(prompt, /"mergeableState": "clean"/);
-  assert.match(prompt, /Do not treat a branch being behind the current base as proof/);
-  assert.match(prompt, /actual three-way merge result/);
-});
-
-test("review context ledger records ordered section budgets", () => {
-  const context = {
-    issue: { number: 123, title: "Sample PR" },
-    comments: [{ author: "alice", body: "Please review this." }],
-    timeline: [{ event: "committed", sha: "abc123" }],
-    previousClawSweeperReview: {
-      status: "found issues before merge.",
-      reviewedSha: "abc123",
-      summary: "Prior review found one blocker.",
-    },
-    relatedItems: [{ number: 122, title: "Related issue" }],
-    pullRequest: { number: 123, additions: 12 },
-    pullFiles: [
-      { filename: "src/example.ts", patch: "line\n".repeat(20) },
-      { filename: "test/example.test.ts", patch: "test\n".repeat(20) },
-    ],
-    pullCommits: [{ sha: "abc123", message: "fix example" }],
-    pullReviewComments: [],
-    counts: {
-      comments: 10,
-      commentsHydrated: 1,
-      commentsTruncated: true,
-      timeline: 1,
-      timelineHydrated: 1,
-      timelineTruncated: false,
-      relatedItems: 1,
-      pullFiles: 120,
-      pullFilesHydrated: 2,
-      pullFilesTruncated: true,
-      pullCommits: 1,
-      pullCommitsHydrated: 1,
-      pullCommitsTruncated: false,
-      pullReviewComments: 0,
-      pullReviewCommentsHydrated: 0,
-      pullReviewCommentsTruncated: false,
-    },
-  };
-
-  const ledger = reviewContextLedgerForTest(context);
-
-  assert.deepEqual(
-    ledger.map(({ section, entries, total, hydrated, truncated }) => [
-      section,
-      entries,
-      total,
-      hydrated,
-      truncated,
-    ]),
-    [
-      ["issue", 1, undefined, undefined, undefined],
-      ["comments", 1, 10, 1, true],
-      ["timeline", 1, 1, 1, false],
-      ["previousClawSweeperReview", 1, undefined, undefined, undefined],
-      ["relatedItems", 1, 1, undefined, undefined],
-      ["pullRequest", 1, undefined, undefined, undefined],
-      ["pullFiles", 2, 120, 2, true],
-      ["pullCommits", 1, 1, 1, false],
-      ["counts", 16, undefined, undefined, undefined],
-    ],
-  );
-  assert.equal(
-    ledger.find((entry) => entry.section === "pullFiles")?.chars,
-    JSON.stringify(context.pullFiles, null, 2).length,
-  );
-  assert.match(
-    renderReviewContextBudgetForTest(context),
-    /- PR files: 2\/120 hydrated, truncated, \d+ chars/,
-  );
-  assert.match(renderReviewContextBudgetForTest(context), /- timeline events: 1\/1 hydrated/);
-  assert.match(renderReviewContextBudgetForTest(context), /- previous ClawSweeper review: 1 entry/);
-});
-
-test("protected labels are normalized and only maintainer-only items stay plannable", () => {
-  assert.deepEqual(protectedLabels(["Security", "bug", "maintainer", "SECURITY"]), [
-    "security",
-    "maintainer",
-  ]);
-  assert.equal(isProtectedItem(item({ labels: ["release-blocker"] })), true);
-  assert.equal(shouldPlanItem(item({ authorAssociation: "MEMBER" })), true);
-  assert.equal(shouldPlanItem(item({ labels: ["maintainer"] })), true);
-  assert.equal(shouldPlanItem(item({ labels: ["maintainer", "security"] })), false);
-  assert.equal(shouldPlanItem(item({ labels: ["beta-blocker"] })), false);
-  assert.equal(shouldPlanItem(item({ labels: ["bug"] })), true);
-});
-
-test("parseGhJson adds gh command context to malformed JSON errors", () => {
-  assert.throws(
-    () => parseGhJson("{", ["api", "repos/openclaw/openclaw/issues"]),
-    /Failed to parse JSON from gh api repos\/openclaw\/openclaw\/issues:/,
-  );
-});
-
-test("parseGhJsonLines adds line number and command context to malformed JSONL errors", () => {
-  assert.throws(
-    () => parseGhJsonLines('{"ok":true}\nnot-json\n', ["issue", "list", "--json", "number"]),
-    /Failed to parse JSON line 2 from gh issue list --json:/,
-  );
-});
-
-test("commit review reports use one canonical path per commit", () => {
-  const sha = "abcdef1234567890abcdef1234567890abcdef12";
-  assert.equal(
-    commitReportRelativePath("openclaw/openclaw", sha),
-    "records/openclaw-openclaw/commits/abcdef1234567890abcdef1234567890abcdef12.md",
-  );
-});
-
-test("commit review parses co-authored-by trailers", () => {
-  assert.deepEqual(
-    parseCoAuthors(`subject
-
-Body text.
-
-Co-authored-by: Alice Example <alice@example.com>
-Co-authored-by: Bob Example <bob@example.com>
-co-authored-by: Alice Example <alice@example.com>
-`),
-    ["Alice Example", "Bob Example"],
-  );
-});
-
-test("commit report since parser accepts compact and natural windows", () => {
-  const now = new Date("2026-04-29T12:00:00.000Z");
-  assert.equal(parseCommitReportSince("6h", now).toISOString(), "2026-04-29T06:00:00.000Z");
-  assert.equal(
-    parseCommitReportSince("24 hours ago", now).toISOString(),
-    "2026-04-28T12:00:00.000Z",
-  );
-  assert.equal(parseCommitReportSince("last 7d", now).toISOString(), "2026-04-22T12:00:00.000Z");
-});
-
-test("skipped non-code commit reports include commit timestamps for listing", () => {
-  const report = skippedNonCodeReport({
-    targetRepo: "openclaw/openclaw",
-    sha: "abcdef1234567890abcdef1234567890abcdef12",
-    metadata: {
-      parents: ["0123456789abcdef0123456789abcdef01234567"],
-      authorName: "Alice",
-      authorEmail: "alice@example.com",
-      committerName: "Bob",
-      committerEmail: "bob@example.com",
-      authoredAt: "2026-04-29T10:00:00Z",
-      committedAt: "2026-04-29T10:05:00Z",
-      coAuthors: [],
-      githubAuthor: "alice",
-      githubCommitter: "bob",
-    },
-    changedFiles: ["docs/usage.md"],
-  });
-  assert.match(report, /commit_authored_at: "2026-04-29T10:00:00Z"/);
-  assert.match(report, /commit_committed_at: "2026-04-29T10:05:00Z"/);
-});
-
-test("commit review cheaply skips documentation-only paths", () => {
-  assert.equal(isReviewableCommitPath("docs/usage.md"), false);
-  assert.equal(isReviewableCommitPath("CHANGELOG.md"), false);
-  assert.equal(isReviewableCommitPath("README.md"), false);
-  assert.equal(isReviewableCommitPath("assets/logo.png"), false);
-  assert.equal(isReviewableCommitPath("test/clawsweeper.test.ts"), true);
-  assert.equal(isReviewableCommitPath("src/clawsweeper.ts"), true);
-  assert.equal(isReviewableCommitPath(".github/workflows/sweep.yml"), true);
-  assert.equal(isReviewableCommitPath("package.json"), true);
-});
-
-test("skipped non-code commit reports still publish green checks", () => {
-  assert.equal(
-    checkConclusionForFrontMatter({ result: "skipped_non_code", highest_severity: "none" }),
-    "success",
-  );
-});
-
-test("commit review check conclusions stay conservative", () => {
-  assert.equal(
-    checkConclusionForFrontMatter({ result: "nothing_found", highest_severity: "none" }),
-    "success",
-  );
-  assert.equal(
-    checkConclusionForFrontMatter({ result: "findings", highest_severity: "high" }),
-    "failure",
-  );
-  assert.equal(
-    checkConclusionForFrontMatter({ result: "findings", highest_severity: "medium" }),
-    "neutral",
-  );
-  assert.equal(checkConclusionForFrontMatter({ result: "inconclusive" }), "neutral");
-});
-
-test("protected labels block close proposals even for otherwise valid decisions", () => {
-  const validation = validateCloseDecision(item({ labels: ["security"] }), closeDecision());
-  assert.equal(validation.ok, false);
-  assert.equal(validation.actionTaken, "skipped_protected_label");
-
-  const action = reviewActionForDecision({
-    item: item({ labels: ["security"] }),
-    decision: closeDecision(),
-    git,
-  });
-  assert.equal(action.actionTaken, "skipped_protected_label");
-  assert.equal(action.closeComment, "");
-});
-
-test("verified fixed maintainer items can become close proposals", () => {
-  const validation = validateCloseDecision(item({ labels: ["maintainer"] }), closeDecision());
-  assert.deepEqual(validation, { ok: true });
-
-  const action = reviewActionForDecision({
-    item: item({ authorAssociation: "MEMBER", labels: ["maintainer"] }),
-    decision: closeDecision(),
-    git,
-  });
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(action.closeComment, /already implemented/);
-});
-
-test("maintainer items stay protected for non-fixed close reasons", () => {
-  const validation = validateCloseDecision(
-    item({ labels: ["maintainer"] }),
-    closeDecision({ closeReason: "duplicate_or_superseded" }),
-  );
-  assert.equal(validation.ok, false);
-  assert.equal(validation.actionTaken, "skipped_protected_label");
-
-  const action = reviewActionForDecision({
-    item: item({ authorAssociation: "MEMBER" }),
-    decision: closeDecision({ closeReason: "duplicate_or_superseded" }),
-    git,
-  });
-  assert.equal(action.actionTaken, "skipped_maintainer_authored");
-  assert.equal(action.closeComment, "");
-});
-
-test("review actions only propose valid closes and never apply directly", () => {
-  const action = reviewActionForDecision({
-    item: item(),
-    decision: closeDecision(),
-    git,
-    runtime: { model: "gpt-5.5", reasoningEffort: "high" },
-  });
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(action.closeComment, /Thanks for the context here/);
-  assert.match(action.closeComment, /shell check/);
-  assert.match(action.closeComment, /already implemented/);
-  assert.match(action.closeComment, /<details>\n<summary>Review details<\/summary>/);
-  assert.match(
-    action.closeComment,
-    /Do we have a high-confidence way to reproduce the issue\?\n\nYes\. Current main can be checked/,
-  );
-  assert.match(
-    action.closeComment,
-    /Is this the best way to solve the issue\?\n\nYes\. Keeping the implementation as-is/,
-  );
-  assert.ok(
-    action.closeComment.indexOf("Is this the best way to solve the issue?") <
-      action.closeComment.indexOf("What I checked:"),
-  );
-  assert.match(action.closeComment, /Likely related people:/);
-  assert.match(action.closeComment, /@alice/);
-  assert.match(action.closeComment, /@bob/);
-  assert.doesNotMatch(action.closeComment, /role: recent maintainer/);
-  assert.match(action.closeComment, /role: recent area contributor/);
-  assert.match(action.closeComment, /Codex review notes: model gpt-5\.5, reasoning high;/);
-});
-
-test("review actions render deterministic close comments when model close comment is empty", () => {
-  const decision = closeDecision({ closeComment: "" });
-  const action = reviewActionForDecision({
-    item: item(),
-    decision,
-    git,
-    runtime: { model: "gpt-5.5", reasoningEffort: "high" },
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(action.closeComment, /Thanks for the context here/);
-  assert.match(action.closeComment, /already implemented/);
-
-  const applyValidation = validateCloseDecision(item(), decision);
-  assert.equal(applyValidation.ok, false);
-  assert.equal(applyValidation.reason, "missing close comment");
-});
-
-test("close comments reference high-confidence merged fixing PRs", () => {
-  const action = reviewActionForDecision({
-    item: item(),
-    decision: closeDecision({
-      fixedPullRequest: {
-        repo: "openclaw/openclaw",
-        number: 456,
-        url: "https://github.com/openclaw/openclaw/pull/456",
-        title: "fix: wire the shell check",
-        mergedAt: "2026-04-28T12:00:00Z",
-        sha: "fedcba9876543210",
-        confidence: "high",
-        source: "GitHub closing PR reference",
-      },
-    }),
-    git,
-    runtime: { model: "gpt-5.5", reasoningEffort: "high" },
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(
-    action.closeComment,
-    /merged PR that appears to have closed this: \[#456: fix: wire the shell check\]\(https:\/\/github\.com\/openclaw\/openclaw\/pull\/456\)/,
-  );
-  assert.match(
-    action.closeComment,
-    /fix evidence: merged PR \[#456\]\(https:\/\/github\.com\/openclaw\/openclaw\/pull\/456\), commit/,
-  );
-});
-
-test("commit PR lookup selects the newest merged pull request", () => {
-  const fixedPullRequest = fixedPullRequestFromCommitPullsForTest([
-    {
-      number: 455,
-      html_url: "https://github.com/openclaw/openclaw/pull/455",
-      title: "fix: older candidate",
-      merged: true,
-      merged_at: "2026-04-27T12:00:00Z",
-      merge_commit_sha: "1111111111111111",
-    },
-    {
-      number: 456,
-      html_url: "https://github.com/openclaw/openclaw/pull/456",
-      title: "fix: wire the shell check",
-      merged_at: "2026-04-28T12:00:00Z",
-      merge_commit_sha: "fedcba9876543210",
-    },
-    {
-      number: 457,
-      html_url: "https://github.com/openclaw/openclaw/pull/457",
-      title: "open follow-up",
-      merged: false,
-    },
-  ]);
-
-  assert.deepEqual(fixedPullRequest, {
-    repo: "openclaw/openclaw",
-    number: 456,
-    url: "https://github.com/openclaw/openclaw/pull/456",
-    title: "fix: wire the shell check",
-    mergedAt: "2026-04-28T12:00:00Z",
-    sha: "fedcba9876543210",
-    confidence: "high",
-    source: "GitHub commit PR lookup",
-  });
-});
-
-test("report-rendered close comments keep merged fixing PR provenance", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "issue",
-      number: "123",
-      title: JSON.stringify("Sample item"),
-      decision: "close",
-      close_reason: "implemented_on_main",
-      action_taken: "proposed_close",
-      fixed_pr_url: "https://github.com/openclaw/openclaw/pull/456",
-      fixed_pr_number: "456",
-      fixed_pr_title: JSON.stringify("fix: wire the shell check"),
-      fixed_pr_merged_at: "2026-04-28T12:00:00Z",
-      fixed_pr_sha: "fedcba9876543210",
-      fixed_pr_confidence: "high",
-      fixed_pr_source: JSON.stringify("GitHub closing PR reference"),
-      fixed_sha: "abcdef1234567890",
-      fixed_at: "2026-04-28T12:00:00Z",
-      main_sha: "abcdef1234567890",
-      review_model: "gpt-5.5",
-      review_reasoning_effort: "high",
-    })}
-
-## Summary
-
-Current main already implements this.
-
-## Best Possible Solution
-
-Keep the implementation as-is.
-
-## Reproduction Assessment
-
-Yes. Current main can be checked by inspecting source and history.
-
-## Solution Assessment
-
-Yes. Keeping the implementation as-is is the narrowest maintainable outcome.
-
-## Evidence
-
-- **implementation:** The feature is present in source.
-  - file: [src/example.ts:12](https://github.com/openclaw/openclaw/blob/abcdef1234567890/src/example.ts#L12)
-  - sha: [abcdef1234567890](https://github.com/openclaw/openclaw/commit/abcdef1234567890)
-
-## Likely Owners
-
-- **@alice:** introduced behavior
-  - reason: git blame points at the fix.
-  - confidence: high
-  - commits: abcdef1234567890
-  - files: src/example.ts
-`,
-    "implemented_on_main",
-  );
-
-  assert.match(
-    comment,
-    /merged PR that appears to have closed this: \[#456: fix: wire the shell check\]\(https:\/\/github\.com\/openclaw\/openclaw\/pull\/456\)/,
-  );
-  assert.match(comment, /fix evidence: merged PR \[#456\]/);
-});
-
-test("close comments suppress duplicate best solution text", () => {
-  const action = reviewActionForDecision({
-    item: item(),
-    decision: closeDecision({
-      summary: "Keep the implementation as-is.",
-      bestSolution: "Keep the implementation as-is.",
-    }),
-    git,
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.doesNotMatch(action.closeComment, /Best possible solution:/);
-});
-
-test("review details show applied AGENTS.md policy status", () => {
-  const action = reviewActionForDecision({
-    item: item(),
-    decision: closeDecision(),
-    git,
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(action.closeComment, /<summary>Review details<\/summary>/);
-  assert.match(action.closeComment, /AGENTS\.md: found and applied where relevant\./);
-});
-
-test("review details show missing AGENTS.md policy status", () => {
-  const action = reviewActionForDecision({
-    item: item(),
-    decision: closeDecision({
-      agentsPolicyStatus: {
-        found: false,
-        readFully: false,
-        applied: false,
-        status: "not_found",
-        summary: "No target repository AGENTS.md was found.",
-      },
-    }),
-    git,
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(action.closeComment, /AGENTS\.md: not found in the target repository\./);
-});
-
-test("likely owner commit links ignore non-sha values", () => {
-  const action = reviewActionForDecision({
-    item: item(),
-    decision: closeDecision({
-      likelyOwners: [
-        {
-          person: "@alice",
-          role: "feature contributor",
-          reason: "The changelog credits a pull request for this feature surface.",
-          commits: ["https://github.com/openclaw/openclaw/pull/76079", " abcdef1234567890 "],
-          files: ["CHANGELOG.md"],
-          confidence: "medium",
-        },
-      ],
-    }),
-    git,
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.doesNotMatch(action.closeComment, /\/commit\/https:/);
-  assert.match(
-    action.closeComment,
-    /\[abcdef123456\]\(https:\/\/github\.com\/openclaw\/openclaw\/commit\/abcdef1234567890\)/,
-  );
-});
-
-test("skill-only OpenClaw PRs can close through ClawHub with upload guidance", () => {
-  const decision = closeDecision({
-    closeReason: "clawhub",
-    summary:
-      "The branch adds an optional bundled skill and does not change required core behavior.",
-    changeSummary: "Adds bundled Higgsfield skill files under skills/higgsfield.",
-    bestSolution:
-      "Publish the skill through ClawHub so it stays installable outside OpenClaw core.",
-    itemCategory: "skill",
-    reproductionStatus: "not_applicable",
-    reproductionConfidence: "high",
-    securityReview: {
-      status: "cleared",
-      summary:
-        "The PR is a skill-only content addition and should move to the community skill path.",
-      concerns: [],
-    },
-    realBehaviorProof: {
-      status: "not_applicable",
-      summary: "Real behavior proof is not needed for a scope-fit close.",
-      evidenceKind: "not_applicable",
-      needsContributorAction: false,
-    },
-  });
-  const pr = item({
-    kind: "pull_request",
-    url: "https://github.com/openclaw/openclaw/pull/78018",
-  });
-
-  assert.equal(validateCloseDecision(pr, decision).ok, true);
-
-  const action = reviewActionForDecision({
-    item: pr,
-    decision,
-    git,
-  });
-
-  assert.equal(action.actionTaken, "proposed_close");
-  assert.match(action.closeComment, /ClawHub\.com/);
-  assert.match(action.closeComment, /upload or publish/i);
-  assert.match(action.closeComment, /installable community skill/);
-});
-
-test("ClawHub policy allows main-implemented issue and PR close proposals", () => {
-  const implementedPr = validateCloseDecision(
-    item({
-      repo: "openclaw/clawhub",
-      kind: "pull_request",
-      url: "https://github.com/openclaw/clawhub/pull/123",
-    }),
-    closeDecision(),
-  );
-  assert.equal(implementedPr.ok, true);
-
-  const implementedIssue = validateCloseDecision(
-    item({
-      repo: "openclaw/clawhub",
-      kind: "issue",
-      url: "https://github.com/openclaw/clawhub/issues/123",
-    }),
-    closeDecision(),
-  );
-  assert.equal(implementedIssue.ok, true);
-
-  const nonImplementedPr = validateCloseDecision(
-    item({
-      repo: "openclaw/clawhub",
-      kind: "pull_request",
-      url: "https://github.com/openclaw/clawhub/pull/123",
-    }),
-    closeDecision({ closeReason: "cannot_reproduce" }),
-  );
-  assert.equal(nonImplementedPr.ok, false);
-  assert.equal(nonImplementedPr.actionTaken, "skipped_invalid_decision");
-});
-
-test("ClawSweeper policy allows self implemented-on-main issue and PR close proposals", () => {
-  const implementedPr = validateCloseDecision(
-    item({
-      repo: "openclaw/clawsweeper",
-      kind: "pull_request",
-      url: "https://github.com/openclaw/clawsweeper/pull/17",
-    }),
-    closeDecision(),
-  );
-  assert.equal(implementedPr.ok, true);
-
-  const implementedIssue = validateCloseDecision(
-    item({
-      repo: "openclaw/clawsweeper",
-      kind: "issue",
-      url: "https://github.com/openclaw/clawsweeper/issues/17",
-    }),
-    closeDecision(),
-  );
-  assert.equal(implementedIssue.ok, true);
-});
-
-function failedReviewReport(overrides = {}) {
-  return `${workPlanCandidateReport({
-    repository: "openclaw/openclaw",
-    number: 4242,
-    type: "pull_request",
-    review_status: "failed",
-    pull_head_sha: "abc123def456",
-    decision: "keep_open",
-    confidence: "low",
-    action_taken: "kept_open",
-    work_candidate: "none",
-    ...overrides,
-  })}
-
-## Summary
-
-Codex review failed: timeout.
-
-## Evidence
-
-- **failure reason:** timeout
-- **codex failure detail:** Codex worker timed out after 600000ms with ETIMEDOUT.
-`;
-}
-
-test("failed review retry eligibility requires infrastructure failure and matching live head", () => {
-  const markdown = failedReviewReport();
-  const now = Date.parse("2026-06-05T20:00:00Z");
-
-  assert.equal(isInfrastructureFailedReviewForTest(markdown), true);
-  assert.deepEqual(
-    failedReviewRetryEligibilityForTest({
-      markdown,
-      liveState: "open",
-      liveHeadSha: "abc123def456",
-      now,
-      maxAttempts: 2,
-      cooldownMs: 45 * 60 * 1000,
-    }),
-    {
-      repo: "openclaw/openclaw",
-      number: 4242,
-      action: "planned_failed_review_retry",
-      reason: "eligible infrastructure failed review at head abc123def456",
-      headSha: "abc123def456",
-      attempts: 0,
-    },
-  );
-  assert.equal(
-    failedReviewRetryEligibilityForTest({
-      markdown,
-      liveState: "open",
-      liveHeadSha: "def456abc123",
-      now,
-      maxAttempts: 2,
-      cooldownMs: 45 * 60 * 1000,
-    }).action,
-    "skipped_stale_head",
-  );
-  assert.equal(
-    failedReviewRetryEligibilityForTest({
-      markdown: failedReviewReport({ review_status: "complete" }),
-      liveState: "open",
-      liveHeadSha: "abc123def456",
-      now,
-      maxAttempts: 2,
-      cooldownMs: 45 * 60 * 1000,
-    }).action,
-    "skipped_not_failed_review",
-  );
-});
-
-test("failed review retry eligibility treats Codex rate limits as infrastructure failures", () => {
-  const markdown = failedReviewReport({
-    repository: "steipete/oracle",
-    number: 250,
-  }).replace(
-    "Codex worker timed out after 600000ms with ETIMEDOUT.",
-    [
-      "stream disconnected: Rate limit reached for hidden-model (for limit test) on tokens per min (TPM). Please try again in 581ms.",
-      "ERROR: The model quoted-model does not exist or you do not have access to it.",
-    ].join("\n"),
-  );
-
-  assert.equal(isInfrastructureFailedReviewForTest(markdown), true);
-});
-
-test("failed review retry eligibility treats model access failures as terminal", () => {
-  const markdown = failedReviewReport({ review_terminal_failure: true })
-    .replaceAll(
-      "Codex review failed: timeout.",
-      "Codex review failed: model unavailable or access denied.",
-    )
-    .replaceAll(
-      "Codex worker timed out after 600000ms with ETIMEDOUT.",
-      [
-        "ERROR: stream disconnected before completion: The model hidden-model does not exist or you do not have access to it.",
-        "- **codex terminal error:** ERROR: stream disconnected before completion: The model hidden-model does not exist or you do not have access to it.",
-      ].join("\n"),
-    );
-
-  assert.equal(isInfrastructureFailedReviewForTest(markdown), false);
-  assert.equal(
-    failedReviewRetryEligibilityForTest({
-      markdown,
-      liveState: "open",
-      liveHeadSha: "abc123def456",
-      now: Date.parse("2026-06-05T20:00:00Z"),
-      maxAttempts: 2,
-      cooldownMs: 45 * 60 * 1000,
-    }).action,
-    "skipped_non_infrastructure_failure",
-  );
-});
-
-test("failed review retry ignores terminal-looking text outside dedicated evidence", () => {
-  const markdown = failedReviewReport().replace(
-    "## Summary",
-    [
-      "Contributor-controlled text: ERROR: The model hidden-model does not exist or you do not have access to it.",
-      "",
-      "## Summary",
-    ].join("\n"),
-  );
-
-  assert.equal(isInfrastructureFailedReviewForTest(markdown), true);
-});
-
-test("failed review retry ignores terminal-looking text injected into rendered evidence", () => {
-  const markdown = failedReviewReport().replace(
-    "Codex worker timed out after 600000ms with ETIMEDOUT.",
-    [
-      "Codex worker timed out after 600000ms with ETIMEDOUT.",
-      "- **codex terminal error:** ERROR: The model fake does not exist or you do not have access to it.",
-    ].join("\n"),
-  );
-
-  assert.equal(isInfrastructureFailedReviewForTest(markdown), true);
-});
-
-test("failed review retry eligibility enforces cooldown and max attempts per head", () => {
-  const now = Date.parse("2026-06-05T20:00:00Z");
-  const recent = failedReviewReport({
-    failed_review_retry_head_sha: "abc123def456",
-    failed_review_retry_count: 1,
-    failed_review_retry_last_at: "2026-06-05T19:30:00Z",
-  });
-  const exhausted = failedReviewReport({
-    failed_review_retry_head_sha: "abc123def456",
-    failed_review_retry_count: 2,
-    failed_review_retry_last_at: "2026-06-05T18:00:00Z",
-  });
-
-  assert.equal(
-    failedReviewRetryEligibilityForTest({
-      markdown: recent,
-      liveState: "open",
-      liveHeadSha: "abc123def456",
-      now,
-      maxAttempts: 2,
-      cooldownMs: 45 * 60 * 1000,
-    }).action,
-    "skipped_retry_cooldown",
-  );
-  assert.equal(
-    failedReviewRetryEligibilityForTest({
-      markdown: exhausted,
-      liveState: "open",
-      liveHeadSha: "abc123def456",
-      now,
-      maxAttempts: 2,
-      cooldownMs: 45 * 60 * 1000,
-    }).action,
-    "skipped_retry_exhausted",
-  );
-});
-
-test("failed review retry exhaustion is idempotent for the same head", () => {
-  const root = mkdtempSync(tmpPrefix);
-  try {
-    const itemsDir = join(root, "items");
-    const reportPath = join(root, "failed-review-retry-report.json");
-    const itemPath = join(itemsDir, "4242.md");
-    mkdirSync(itemsDir, { recursive: true });
-    writeFileSync(
-      itemPath,
-      failedReviewReport({
-        failed_review_retry_head_sha: "abc123def456",
-        failed_review_retry_count: 2,
-        failed_review_retry_last_at: "2026-06-05T18:00:00Z",
-      }),
-      "utf8",
-    );
-
-    const ghMock = `#!/usr/bin/env node
-const args = process.argv.slice(2);
-const path = args.find((arg) => arg.startsWith("repos/")) || "";
-if (path.endsWith("/issues/4242")) {
-  console.log(JSON.stringify({
-    number: 4242,
-    title: "Failed review retry sample",
-    html_url: "https://github.com/openclaw/openclaw/pull/4242",
-    created_at: "2026-06-01T00:00:00Z",
-    updated_at: "2026-06-01T01:00:00Z",
-    closed_at: null,
-    state: "open",
-    locked: false,
-    active_lock_reason: null,
-    author_association: "CONTRIBUTOR",
-    user: { login: "contributor" },
-    labels: [],
-    pull_request: {}
-  }));
-  process.exit(0);
-}
-if (path.endsWith("/pulls/4242")) {
-  console.log("abc123def456");
-  process.exit(0);
-}
-console.error("unexpected gh args: " + args.join(" "));
-process.exit(1);
-`;
-
-    const runRetry = () => {
-      execFileSync(process.execPath, [
-        "dist/clawsweeper.js",
-        "retry-failed-reviews",
-        "--target-repo",
-        "openclaw/openclaw",
-        "--items-dir",
-        itemsDir,
-        "--item-number",
-        "4242",
-        "--max-attempts",
-        "2",
-        "--cooldown-minutes",
-        "45",
-        "--report-path",
-        reportPath,
-      ]);
-    };
-
-    withMockGh(root, ghMock, () => {
-      runRetry();
-      const afterFirstRun = readFileSync(itemPath, "utf8");
-      assert.match(afterFirstRun, /^failed_review_retry_status: exhausted$/m);
-      assert.equal((afterFirstRun.match(/^## Failed Review Retry$/gm) ?? []).length, 1);
-
-      runRetry();
-      const afterSecondRun = readFileSync(itemPath, "utf8");
-      assert.equal(afterSecondRun, afterFirstRun);
-    });
-
-    const report = JSON.parse(readFileSync(reportPath, "utf8")) as Array<{
-      action: string;
-      number: number;
-    }>;
-    assert.deepEqual(report, [
-      {
-        repo: "openclaw/openclaw",
-        number: 4242,
-        action: "skipped_retry_already_exhausted",
-        reason: "retry attempts exhausted for head abc123def456: 2/2",
-        headSha: "abc123def456",
-        attempts: 2,
-        reportPath: itemPath,
-      },
-    ]);
-  } finally {
-    rmSync(root, { recursive: true, force: true });
-  }
-});
 
 function unconfirmedProductDirectionCloseReport(overrides = {}) {
   return `${workPlanCandidateReport({
@@ -3500,918 +2466,6 @@ if (args[0] === "api" && /\\/issues\\/321\\/comments(?:\\?|$)/.test(path)) {
   } finally {
     rmSync(root, { recursive: true, force: true });
   }
-});
-
-test("security-needs-attention reports block unopted repair and automerge pass markers", () => {
-  const securitySection = `
-## Security Review
-
-Status: needs_attention
-
-Summary: The patch exposes a broader token scope and needs maintainer security review.
-
-Concerns:
-
-- **[high] Avoid broad token reuse:** \`src/auth/token.ts:42\`
-  - body: The patch can reuse a token with broader scopes than the caller requested.
-  - confidence: 0.91
-`;
-  const repairMarkers = reviewAutomationMarkersFromReport(`${reportFrontMatter({
-    type: "pull_request",
-    number: "74123",
-    pull_head_sha: "abc123def456",
-    decision: "keep_open",
-    confidence: "high",
-    work_candidate: "queue_fix_pr",
-  })}
-
-## Summary
-
-Needs a repair.
-
-${securitySection}
-`);
-
-  assert.match(repairMarkers, /clawsweeper-security:security-sensitive/);
-  assert.match(repairMarkers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(repairMarkers, /clawsweeper-verdict:needs-changes/);
-  assert.doesNotMatch(repairMarkers, /clawsweeper-action:fix-required/);
-
-  const autofixRepairMarkers = reviewAutomationMarkersFromReport(`${reportFrontMatter({
-    type: "pull_request",
-    number: "74125",
-    pull_head_sha: "abc789def123",
-    decision: "keep_open",
-    confidence: "high",
-    labels: JSON.stringify(["clawsweeper:autofix"]),
-    work_candidate: "queue_fix_pr",
-  })}
-
-## Summary
-
-Needs an opted-in repair.
-
-${securitySection}
-`);
-
-  assert.match(autofixRepairMarkers, /clawsweeper-security:security-sensitive/);
-  assert.match(autofixRepairMarkers, /clawsweeper-verdict:needs-changes/);
-  assert.match(autofixRepairMarkers, /clawsweeper-action:fix-required/);
-  assert.match(autofixRepairMarkers, /finding=security-review/);
-  assert.doesNotMatch(autofixRepairMarkers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(autofixRepairMarkers, /clawsweeper-verdict:pass/);
-
-  const automergeMarkers = reviewAutomationMarkersFromReport(`${reportFrontMatter({
-    type: "pull_request",
-    number: "74124",
-    pull_head_sha: "def456abc123",
-    decision: "keep_open",
-    confidence: "high",
-    review_status: "complete",
-    labels: JSON.stringify(["clawsweeper:automerge"]),
-    work_candidate: "none",
-  })}
-
-## Summary
-
-Would otherwise pass automerge.
-
-${securitySection}
-`);
-
-  assert.match(automergeMarkers, /clawsweeper-security:security-sensitive/);
-  assert.match(automergeMarkers, /clawsweeper-verdict:needs-changes/);
-  assert.match(automergeMarkers, /clawsweeper-action:fix-required/);
-  assert.match(automergeMarkers, /finding=security-review/);
-  assert.doesNotMatch(automergeMarkers, /clawsweeper-verdict:pass/);
-  assert.doesNotMatch(automergeMarkers, /clawsweeper-verdict:needs-human/);
-});
-
-test("pull request keep-open review comments suppress duplicate remaining risk text", () => {
-  const duplicateRisk = "Run the automerge smoke after the repair lane is green.";
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74267",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this smoke-test PR open for maintainer review.
-
-## What This Changes
-
-Adds regression coverage for automerge repair smoke comments.
-
-## Risks / Open Questions
-
-${duplicateRisk}
-
-## Work Candidate
-
-Candidate: none
-
-Confidence: low
-
-Priority: low
-
-Status: none
-
-Reason: ${duplicateRisk}
-`,
-    "none",
-  );
-
-  assert.ok(comment.includes(`**Next step before merge**\n- [P2] ${duplicateRisk}`));
-  assert.doesNotMatch(comment, /Remaining risk \/ open question:/);
-  assert.doesNotMatch(comment, /\*\*Risk before merge\*\*/);
-  assert.equal(comment.split(duplicateRisk).length - 1, 1);
-});
-
-test("pull request keep-open review comments prefix each merge risk bullet", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74269",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this multi-risk PR open for maintainer review.
-
-## What This Changes
-
-Changes generated review-comment formatting.
-
-## Best Possible Solution
-
-Confirm both merge risks before merge.
-
-## Risks / Open Questions
-
-- Blocked workflow actions must render as P1.
-- Timeout fallback wording should remain scannable.
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Risk before merge\*\*/);
-  assert.match(comment, /- \[P1\] Blocked workflow actions must render as P1\./);
-  assert.match(comment, /- \[P2\] Timeout fallback wording should remain scannable\./);
-  assert.doesNotMatch(comment, /- \[P1\] Blocked workflow actions.*\n- Timeout fallback/s);
-});
-
-test("pull request risk text does not priority-prefix routine CI noise", () => {
-  const routineCiRisk = "CI checks are red on this branch and may be unrelated to the diff.";
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74269",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-    })}
-
-## Summary
-
-Keep this PR open while maintainers verify check state.
-
-## What This Changes
-
-Updates review guidance.
-
-## Best Possible Solution
-
-Merge after the unrelated CI state is understood.
-
-## Risks / Open Questions
-
-${routineCiRisk}
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Risk before merge\*\*/);
-  assert.match(comment, new RegExp(`- ${routineCiRisk.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`));
-  assert.doesNotMatch(
-    comment,
-    new RegExp(`\\[P[12]\\] ${routineCiRisk.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-  );
-});
-
-test("pull request next step does not priority-prefix routine required status checks", () => {
-  const routineStatusSteps = [
-    "Merge after required status checks are green.",
-    "Merge after required checks pass.",
-    "Merge after status checks pass.",
-    "Merge once required checks have passed.",
-    "Wait for status checks to pass.",
-    "Merge after required checks.",
-    "Wait for required checks.",
-    "Merge after required checks pass and no failures are seen.",
-    "Merge after required checks pass without failures.",
-    "Merge after required checks pass without any failures.",
-    "CI checks pass without test failures.",
-    "Merge after required checks pass without any test failures.",
-    "CI checks pass but no failures are seen.",
-    "CI checks pass but maintainer review is still required.",
-    "Required checks pass and required approvals are complete.",
-    "CI checks are red but may pass on rerun.",
-    "Merge after required checks and maintainer review.",
-  ];
-  for (const routineStatusStep of routineStatusSteps) {
-    const comment = renderReviewCommentFromReport(
-      `${reportFrontMatter({
-        type: "pull_request",
-        number: "74273",
-        decision: "keep_open",
-        close_reason: "none",
-        work_candidate: "none",
-        pull_head_sha: "abc123def460",
-      })}
-
-## Summary
-
-Keep this PR open until normal merge gates pass.
-
-## What This Changes
-
-Updates review guidance.
-
-## Best Possible Solution
-
-${routineStatusStep}
-`,
-      "none",
-    );
-
-    assert.match(comment, /\*\*Next step before merge\*\*/);
-    assert.match(
-      comment,
-      new RegExp(`- ${routineStatusStep.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-    );
-    assert.doesNotMatch(
-      comment,
-      new RegExp(`\\[P[12]\\] ${routineStatusStep.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-    );
-  }
-});
-
-test("pull request risk text keeps diff-caused CI risk actionable", () => {
-  const actionableCiRisk = "The workflow change could cause CI checks to fail after merge.";
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74270",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def457",
-    })}
-
-## Summary
-
-Keep this PR open while maintainers verify workflow behavior.
-
-## What This Changes
-
-Updates workflow handling.
-
-## Best Possible Solution
-
-Merge after the workflow risk is addressed.
-
-## Risks / Open Questions
-
-${actionableCiRisk}
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Risk before merge\*\*/);
-  assert.match(
-    comment,
-    new RegExp(`- \\[P1\\] ${actionableCiRisk.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-  );
-});
-
-test("pull request risk text keeps diff-caused status-check risk actionable", () => {
-  const actionableStatusRisk = "The workflow change could cause status checks to fail after merge.";
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74271",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def458",
-    })}
-
-## Summary
-
-Keep this PR open while maintainers verify workflow behavior.
-
-## What This Changes
-
-Updates workflow handling.
-
-## Best Possible Solution
-
-Merge after the status-check risk is addressed.
-
-## Risks / Open Questions
-
-${actionableStatusRisk}
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Risk before merge\*\*/);
-  assert.match(
-    comment,
-    new RegExp(`- \\[P1\\] ${actionableStatusRisk.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-  );
-});
-
-test("pull request risk text keeps diff-caused required-check risk actionable", () => {
-  const actionableRequiredRisk =
-    "The workflow change could cause required checks to fail after merge.";
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "74272",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def459",
-    })}
-
-## Summary
-
-Keep this PR open while maintainers verify workflow behavior.
-
-## What This Changes
-
-Updates workflow handling.
-
-## Best Possible Solution
-
-Merge after the required-check risk is addressed.
-
-## Risks / Open Questions
-
-${actionableRequiredRisk}
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Risk before merge\*\*/);
-  assert.match(
-    comment,
-    new RegExp(`- \\[P1\\] ${actionableRequiredRisk.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-  );
-});
-
-test("pull request risk text keeps broken passing-check risk actionable", () => {
-  const actionablePassingRisks = [
-    "The workflow change makes required checks pass even when tests fail.",
-    "CI checks are passing despite tests failing.",
-    "Security exposure remains even though status checks are green.",
-    "CI checks are green but snapshot drift blocks merge.",
-    "CI checks are green but the app crashes on startup.",
-    "Status checks pass despite data loss.",
-    "CI checks pass without running tests for the changed path.",
-    "CI checks pass with tests disabled.",
-    "Required checks pass after skipping the changed-path tests.",
-    "CI checks pass without failures, but required docs are missing.",
-    "CI checks pass and tests fail.",
-    "Required checks pass and required docs are missing.",
-    "Required checks pass and required approvals are complete, but required docs are missing.",
-    "CI checks pass and required approvals are complete, but coverage is too low.",
-    "CI checks pass because tests are mock-only.",
-    "Status checks pass because validation is stubbed.",
-    "CI checks pass but maintainer review is still required because tests were skipped.",
-    "CI checks pass and required approvals are complete, but tests are disabled.",
-    "CI checks pass with no tests for the changed path.",
-    "CI checks are green with no validation.",
-    "CI checks pass with only mocked tests.",
-    "CI checks pass with insufficient coverage.",
-    "CI checks pass and no tests run for this path.",
-    "CI checks pass and no validation runs.",
-    "CI checks pass and do not run tests for the changed path.",
-    "CI checks pass and tests do not cover the changed path.",
-    "CI checks pass and the changed path is untested.",
-    "CI checks pass and a manual data migration is required before merge.",
-  ];
-  for (const actionablePassingRisk of actionablePassingRisks) {
-    const comment = renderReviewCommentFromReport(
-      `${reportFrontMatter({
-        type: "pull_request",
-        number: "74274",
-        decision: "keep_open",
-        close_reason: "none",
-        work_candidate: "none",
-        pull_head_sha: "abc123def461",
-      })}
-
-## Summary
-
-Keep this PR open while maintainers verify workflow behavior.
-
-## What This Changes
-
-Updates workflow handling.
-
-## Best Possible Solution
-
-Merge after the required-check risk is addressed.
-
-## Risks / Open Questions
-
-${actionablePassingRisk}
-`,
-      "none",
-    );
-
-    assert.match(comment, /\*\*Risk before merge\*\*/);
-    assert.match(
-      comment,
-      new RegExp(`- \\[P[01]\\] ${actionablePassingRisk.replace(/[.*+?^${}()|[\]\\]/g, "\\$&")}`),
-    );
-  }
-});
-
-test("OpenClaw pull request comments render PR surface inside evidence details", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      repository: "openclaw/openclaw",
-      type: "pull_request",
-      number: "12345",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pr_surface_files: JSON.stringify([
-        { path: "src/runtime.ts", additions: 10, deletions: 2 },
-        { path: "src/runtime.test.ts", additions: 7, deletions: 1 },
-        { path: "docs/usage.md", additions: 4, deletions: 0 },
-      ]),
-      pr_surface_files_truncated: "false",
-      review_metrics: JSON.stringify([]),
-    })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Adds a small runtime change with tests and docs.
-`,
-    "none",
-  );
-
-  const evidenceDetails = detailsBody(comment, "Evidence reviewed");
-  const visibleBeforeEvidence = comment.slice(
-    0,
-    comment.indexOf("<summary>Evidence reviewed</summary>"),
-  );
-
-  assert.match(
-    visibleBeforeEvidence,
-    /PR surface: Source \+8, Tests \+6, Docs \+4\. Total \+18 across 3 files\./,
-  );
-  assert.doesNotMatch(visibleBeforeEvidence, /<summary>View PR surface stats<\/summary>/);
-  assert.doesNotMatch(
-    visibleBeforeEvidence,
-    /\| \*\*Total\*\* \| \*\*3\*\* \| \*\*21\*\* \| \*\*3\*\* \| \*\*\+18\*\* \|/,
-  );
-  assert.match(
-    evidenceDetails,
-    /PR surface:\n\nSource \+8, Tests \+6, Docs \+4\. Total \+18 across 3 files\./,
-  );
-  assert.match(evidenceDetails, /<summary>View PR surface stats<\/summary>/);
-  assert.match(
-    evidenceDetails,
-    /\| \*\*Total\*\* \| \*\*3\*\* \| \*\*21\*\* \| \*\*3\*\* \| \*\*\+18\*\* \|/,
-  );
-  assert.match(comment, /\*\*Review metrics:\*\* none identified\./);
-  assert.ok(comment.indexOf("PR surface:") < comment.indexOf("**Review metrics:**"));
-  assert.ok(comment.indexOf("**Review metrics:**") < comment.indexOf("**Merge readiness**"));
-});
-
-test("pull request comments render one review metric digest item", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      repository: "openclaw/openclaw",
-      type: "pull_request",
-      number: "12345",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      review_metrics: JSON.stringify([
-        {
-          label: "Workflow surfaces changed",
-          value: "1 workflow changed",
-          reason:
-            "The PR changes repository automation behavior that maintainers should review before merge.",
-        },
-      ]),
-    })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Updates repository automation.
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Review metrics:\*\* 1 noteworthy metric\./);
-  assert.match(
-    comment,
-    /- \*\*Workflow surfaces changed:\*\* 1 workflow changed\. The PR changes repository automation behavior that maintainers should review before merge\./,
-  );
-});
-
-test("pull request comments render multiple review metric digest items near PR surface", () => {
-  const comment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      repository: "openclaw/openclaw",
-      type: "pull_request",
-      number: "12345",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pr_surface_files: JSON.stringify([{ path: "src/runtime.ts", additions: 10, deletions: 2 }]),
-      pr_surface_files_truncated: "false",
-      review_metrics: JSON.stringify([
-        {
-          label: "Config/default surfaces changed",
-          value: "2 added, 1 changed, 0 removed",
-          reason:
-            "The PR introduces user-facing configuration behavior that maintainers should review before merge.",
-        },
-        {
-          label: "Proof files affected",
-          value: "3 files affected",
-          reason:
-            "The PR touches proof-related code where green unit tests do not cover every runtime path.",
-        },
-      ]),
-    })}
-
-## Summary
-
-Keep this PR open for maintainer review.
-
-## What This Changes
-
-Adds configuration behavior and proof updates.
-`,
-    "none",
-  );
-
-  assert.match(comment, /\*\*Review metrics:\*\* 2 noteworthy metrics\./);
-  assert.match(
-    comment,
-    /- \*\*Config\/default surfaces changed:\*\* 2 added, 1 changed, 0 removed\./,
-  );
-  assert.match(comment, /- \*\*Proof files affected:\*\* 3 files affected\./);
-  assert.ok(comment.indexOf("PR surface:") < comment.indexOf("**Review metrics:**"));
-  assert.ok(comment.indexOf("**Review metrics:**") < comment.indexOf("**Merge readiness**"));
-});
-
-test("PR surface is OpenClaw pull-request only", () => {
-  const frontMatter = {
-    decision: "keep_open",
-    close_reason: "none",
-    work_candidate: "none",
-    pr_surface_files: JSON.stringify([{ path: "src/runtime.ts", additions: 10, deletions: 2 }]),
-    pr_surface_files_truncated: "false",
-  };
-  const body = `
-
-## Summary
-
-Keep this open.
-`;
-
-  const otherRepoComment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      ...frontMatter,
-      repository: "example/project",
-      type: "pull_request",
-    })}${body}`,
-    "none",
-  );
-  const issueComment = renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      ...frontMatter,
-      repository: "openclaw/openclaw",
-      type: "issue",
-    })}${body}`,
-    "none",
-  );
-
-  assert.doesNotMatch(otherRepoComment, /PR surface:/);
-  assert.doesNotMatch(issueComment, /PR surface:/);
-});
-
-function mergeRiskReviewComment({
-  risk,
-  options,
-  bestSolution = "Resolve the merge risk before maintainers decide whether to land this PR.",
-}: {
-  risk: string;
-  options: readonly Record<string, unknown>[];
-  bestSolution?: string;
-}): string {
-  return renderReviewCommentFromReport(
-    `${reportFrontMatter({
-      type: "pull_request",
-      number: "83400",
-      decision: "keep_open",
-      close_reason: "none",
-      work_candidate: "none",
-      pull_head_sha: "abc123def456",
-      merge_risk_options: JSON.stringify(options),
-    })}
-
-## Summary
-
-Keep this fail-closed provider-routing PR open for maintainer review.
-
-## What This Changes
-
-Changes missing Codex harness selection from fallback-tolerant behavior to a typed fail-closed error.
-
-## Best Possible Solution
-
-${bestSolution}
-
-## Risks / Open Questions
-
-${risk}
-
-## Work Candidate
-
-Candidate: none
-
-Confidence: low
-
-Priority: low
-
-Status: none
-
-Reason: Confirm whether this intentional fail-closed behavior is acceptable for existing fallback users.
-`,
-    "none",
-  );
-}
-
-test("pull request keep-open review comments render repairable merge-risk options with one copy block", () => {
-  const mergeRisk =
-    "Existing users configured with a missing Codex harness would fail closed instead of continuing through their fallback model.";
-  const comment = mergeRiskReviewComment({
-    risk: mergeRisk,
-    bestSolution:
-      "Keep fallback behavior as the default and add a strict config option for the fail-closed behavior.",
-    options: [
-      {
-        title: "Preserve existing behavior by default",
-        body: "Keep fallback behavior as the default and add a strict config option for the fail-closed behavior.",
-        category: "fix_before_merge",
-        recommended: true,
-        automergeInstruction:
-          "Keep fallback behavior as the default and add a strict config option for the fail-closed behavior.",
-      },
-      {
-        title: "Make the breaking change explicit",
-        body: "Keep fail-closed behavior only if docs, tests, and release notes warn existing fallback users.",
-        category: "fix_before_merge",
-        recommended: false,
-        automergeInstruction: "",
-      },
-      {
-        title: "Do not merge as-is",
-        body: "Pause or close this PR if maintainers do not want to take this compatibility risk.",
-        category: "pause_or_close",
-        recommended: false,
-        automergeInstruction: "",
-      },
-    ],
-  });
-
-  assert.match(comment, /\*\*Risk before merge\*\*/);
-  assert.match(comment, new RegExp(escapeRegExpForTest(mergeRisk)));
-  assert.doesNotMatch(comment, /Why this matters:/);
-  assert.match(
-    comment,
-    /\*\*Maintainer options:\*\*\n1\. \*\*Preserve existing behavior by default \(recommended\)\*\*/,
-  );
-  assert.match(comment, /2\. \*\*Make the breaking change explicit\*\*/);
-  assert.match(comment, /3\. \*\*Do not merge as-is\*\*/);
-  assert.match(
-    comment,
-    /<summary>Copy recommended automerge instruction<\/summary>[\s\S]*@clawsweeper automerge\n\nSpecial instructions:\nKeep fallback behavior as the default and add a strict config option for the fail-closed behavior\./,
-  );
-  assert.doesNotMatch(comment, /Remaining risk \/ open question:/);
-});
-
-test("pull request keep-open review comments strip nested ClawSweeper commands from copy block", () => {
-  const comment = mergeRiskReviewComment({
-    risk: "Delivery repair should not run with nested bot commands in the pasteable instruction.",
-    bestSolution: "Repair duplicate delivery and add regression coverage before merge.",
-    options: [
-      {
-        title: "Repair delivery before merge",
-        body: "Fix duplicate active-requester delivery and add regression coverage before merge.",
-        category: "fix_before_merge",
-        recommended: true,
-        automergeInstruction:
-          "@clawsweeper autofix this PR: prevent duplicate active-requester delivery and add focused regression coverage before merging.",
-      },
-    ],
-  });
-
-  assert.match(
-    comment,
-    /@clawsweeper automerge\n\nSpecial instructions:\nprevent duplicate active-requester delivery and add focused regression coverage before merging\./,
-  );
-  assert.doesNotMatch(comment, /Special instructions: @clawsweeper/);
-  assert.doesNotMatch(comment, /autofix this PR:/);
-});
-
-test("pull request keep-open review comments can recommend accepting intentional risk without a copy block", () => {
-  const comment = mergeRiskReviewComment({
-    risk: "This hardening intentionally rejects requests that older integrations currently pass.",
-    bestSolution:
-      "Merge only if maintainers accept the compatibility break as intentional hardening.",
-    options: [
-      {
-        title: "Accept the behavior change explicitly",
-        body: "Merge only if maintainers agree the security hardening is worth the compatibility break.",
-        category: "accept_risk",
-        recommended: true,
-        automergeInstruction: "",
-      },
-      {
-        title: "Add migration guidance before merge",
-        body: "Document the rejected request shape and add release-note guidance for affected integrations.",
-        category: "fix_before_merge",
-        recommended: false,
-        automergeInstruction: "",
-      },
-    ],
-  });
-
-  assert.match(comment, /1\. \*\*Accept the behavior change explicitly \(recommended\)\*\*/);
-  assert.doesNotMatch(comment, /Copy recommended ClawSweeper instruction/);
-});
-
-test("pull request keep-open review comments do not force a recommendation for unclear merge risk", () => {
-  const comment = mergeRiskReviewComment({
-    risk: "The PR changes session ownership without proving how existing resumed sessions transition.",
-    options: [
-      {
-        title: "Require a maintainer design decision",
-        body: "Decide whether resumed sessions should migrate, fail fast, or continue using the old ownership model.",
-        category: "pause_or_close",
-        recommended: false,
-        automergeInstruction: "",
-      },
-      {
-        title: "Add migration proof before merge",
-        body: "Add tests or manual validation covering sessions created before this change.",
-        category: "fix_before_merge",
-        recommended: false,
-        automergeInstruction: "",
-      },
-    ],
-  });
-
-  assert.doesNotMatch(comment, /\(recommended\)/);
-  assert.doesNotMatch(comment, /Copy recommended ClawSweeper instruction/);
-});
-
-test("pull request keep-open review comments allow multiple fix-before-merge options", () => {
-  const comment = mergeRiskReviewComment({
-    risk: "The retry path may duplicate queued user messages after partial provider sends.",
-    bestSolution: "Guard retries with delivery state before merge.",
-    options: [
-      {
-        title: "Guard retries with delivery state",
-        body: "Track whether the user message was already sent before retrying provider fallback.",
-        category: "fix_before_merge",
-        recommended: true,
-        automergeInstruction:
-          "Track whether the user message was already sent before retrying provider fallback.",
-      },
-      {
-        title: "Disable fallback after partial sends",
-        body: "Fail fast once delivery starts instead of retrying through another provider.",
-        category: "fix_before_merge",
-        recommended: false,
-        automergeInstruction: "",
-      },
-    ],
-  });
-
-  assert.match(comment, /1\. \*\*Guard retries with delivery state \(recommended\)\*\*/);
-  assert.match(comment, /2\. \*\*Disable fallback after partial sends\*\*/);
-  assert.match(comment, /@clawsweeper automerge/);
-});
-
-test("pull request keep-open review comments include pause or close when risk may outweigh value", () => {
-  const comment = mergeRiskReviewComment({
-    risk: "The PR changes automation proof capture without proving failed paths still upload artifacts.",
-    options: [
-      {
-        title: "Prove artifact parity before merge",
-        body: "Show that artifacts upload on success, failure, and skipped-review paths.",
-        category: "fix_before_merge",
-        recommended: false,
-        automergeInstruction: "",
-      },
-      {
-        title: "Pause or close",
-        body: "Close this PR if maintainers decide the proof-capture regression risk outweighs the workflow cleanup.",
-        category: "pause_or_close",
-        recommended: false,
-        automergeInstruction: "",
-      },
-    ],
-  });
-
-  assert.match(
-    comment,
-    /2\. \*\*Pause or close\*\*  \n   Close this PR if maintainers decide the proof-capture regression risk outweighs the workflow cleanup\./,
-  );
-  assert.doesNotMatch(comment, /Copy recommended ClawSweeper instruction/);
-});
-
-function escapeRegExpForTest(value: string): string {
-  return value.replace(/[.*+?^${}()|[\]\\]/g, "\\$&");
-}
-
-test("pull request review reports carry verdict and repair markers", () => {
-  const markdown = `${reportFrontMatter({
-    type: "pull_request",
-    number: "74065",
-    pull_head_sha: "abc123def456",
-    decision: "keep_open",
-    confidence: "high",
-    work_candidate: "queue_fix_pr",
-  })}
-
-## Summary
-
-Needs one more repair.
-`;
-
-  const markers = reviewAutomationMarkersFromReport(markdown);
-  assert.match(markers, /clawsweeper-verdict:needs-changes/);
-  assert.match(markers, /clawsweeper-action:fix-required/);
-  assert.match(markers, /item=74065/);
-  assert.match(markers, /sha=abc123def456/);
-});
-
-test("pull request reports without a repair candidate pause for human review", () => {
-  const markers = reviewAutomationMarkersFromReport(`${reportFrontMatter({
-    type: "pull_request",
-    number: "74105",
-    pull_head_sha: "abc123def456",
-    decision: "keep_open",
-    confidence: "high",
-    work_candidate: "none",
-  })}
-
-## Summary
-
-Needs maintainer review.
-`);
-
-  assert.match(markers, /clawsweeper-verdict:needs-human/);
-  assert.doesNotMatch(markers, /clawsweeper-verdict:needs-changes/);
-  assert.doesNotMatch(markers, /clawsweeper-action:fix-required/);
-  assert.match(markers, /item=74105/);
-  assert.match(markers, /sha=abc123def456/);
-});
-
-test("non-PR review reports do not carry repair markers", () => {
-  assert.equal(reviewAutomationMarkersFromReport(reportFrontMatter({ type: "issue" })), "");
 });
 
 test("item number args merge and sort workflow inputs", () => {
